app/requirements.txt
Flask
psycopg2-binary
python-dotenv
requests
beautifulsoup4
